chore: split noOverrideDsComponent into separate file
